Added ESlint newline rules; File renames
