+++ public/scripts/editor/serialize-worker.js
